Implement python test cases using checks modules
